Adding support for value objects
