+++ lib/aggregates/value_object.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Aggregates
+  class ValueObject < DomainObject
+  end
+end
